feat: Add invitation system for teams
Introduces a new invitation system that allows users to invite players to their teams and manage these invitations. This includes updating the UI for displaying invitations in the header and the social section, as well as backend support for sending, accepting, rejecting, and canceling invitations.

This change also refactors some type definitions and updates the notification sound to be self-contained within the application.