feat(social): Add 'Available Today' section
Introduces a new section in the social view for users to see which players are available to play today. This includes updates to the `SocialSection` type and the addition of a new view component for displaying this information. Player types have also been extended with `isAvailableToday` and `lastKnownLocation` fields to support this feature.